feat: 进程内 Gopeed 引擎、公共 Movies 落盘、下载进度条与 Tinker 补丁入口
- 引擎: Gopeed 改为 gomobile 进程内绑定(app/libs/libgopeed.aar), 移除 assets/gopeed/gopeed-arm64, 新增 GopeedTransport/GopeedTaskClient/GopeedEngine
- 修复: 登录 mid、cid 64 位溢出、JsonNull 解析、CDN 403(浏览器 UA + Referer)、Android 14 FGS dataSync
- 校验: PlaybackVerifier 增加 selectTrack 与 KEY_MAX_INPUT_SIZE 缓冲, 修复合并产物被误判无法读取样本
- 落盘: 成品写公共 Movies/BiliFolderDownloader/, 分片与 zip 留应用专属目录, 成功后扫描媒体库
- 进度: 新增当前文件字节级进度条(下载页 + 前台通知)
- 打包: 默认不打包; 仅开启 WebDAV 自动上传时打包上传, 远端确认后删除本地 zip
- 调试: 外置 WebDAV 凭据(debug 源集)与调试日志 db 上传按钮
- 补丁: 设置页新增 Tinker 补丁包选择入口
- 版本: 1.0.14 (versionCode 15)

+++ app/src/main/java/com/bilifolder/downloader/data/model/Models.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 data class VideoInfo(
     val bvid: String,
     val title: String,
-    val cid: Int,
+    val cid: Long,
     val avid: Long = 0,
     val page: Int = 1,
 )
